+++ sync/SyncDataCollection.js
@@ -34,9 +34,10 @@ var SyncDataCollection = (function () {
         this.lastModifiedPropName = lastModifiedPropName;
     }
     /** Sync down a set of data collections
+     * @param <R> the sync down error type
      * @param params: parameters to pass to the sync function
      * @param syncSettingsAry: an array of SyncSettings object to sync
-     * @param [clearData=true]: true to clear the existing psData collection before syncing, false to leave it as-is
+     * @param syncDownOp the type of sync to perform
      * @return a map of {@code syncSettingsAry} collection names to the promises that will complete when they finish syncing
      */
     SyncDataCollection.prototype.syncDownCollections = function (params, syncSettingsAry, syncDownOp) {
@@ -55,14 +56,17 @@ var SyncDataCollection = (function () {
      * @param <E> the local collection data model. This type should contain deleted, synched, and last modified properties corresponding to the prop names passed to the constructor
      * @param <F> the local collection data model with optional properties. This type should contain deleted, synched, and last modified properties corresponding to the prop names passed to the constructor
      * @param <S> the remote data model. This type should contain deleted, synched, and last modified properties corresponding to the prop names passed to the constructor
+     * @param <R1> the sync down function error type
+     * @param <R2> the process results callback error type
      */
-    SyncDataCollection.prototype.syncDownCollection = function (params, table, syncDownFunc, processResultItemsCallback) {
+    SyncDataCollection.prototype.syncDownCollection = function (params, table, syncDownFunc, processResultsCallback) {
         var self = this;
         var dfd = Defer.newDefer();
-        function syncFailure(msg) {
+        function syncFailure(err) {
             dfd.reject({
                 collectionName: table.getName(),
-                error: msg
+                syncingDown: true,
+                error: err,
             });
         }
         function saveData() {
@@ -71,7 +75,7 @@ var SyncDataCollection = (function () {
             dfd.resolve(null);
         }
         syncDownFunc(params).done(function (items) {
-            var promise = processResultItemsCallback(items);
+            var promise = processResultsCallback(items);
             if (promise != null && promise["then"]) {
                 promise.then(saveData, syncFailure);
             }
@@ -256,7 +260,7 @@ var SyncDataCollection = (function () {
                 resultItems.push(itemConverter(item));
             }
             else {
-                console.error("Error syncing " + collName);
+                throw new Error("Cannot syncing '" + collName + "' collection without primary keys");
             }
         }
         return resultItems;
@@ -274,7 +278,7 @@ var SyncDataCollection = (function () {
                 resultItems.push(itemConverter(item));
             }
             else {
-                console.error("Error syncing " + collName);
+                throw new Error("Cannot sync '" + collName + "' collection without primary keys");
             }
         }
         return resultItems;